backend/requirements.txt
fastapi
uvicorn[standard]
langchain-groq
langchain
langchain-community
pypdf
chromadb
mysql-connector-python
python-dotenv
sqlalchemy
